+++ src/main/java/com/notenoughmail/kubejs_tfc/util/EventHandlers.java
@@ -5,6 +5,7 @@ import com.notenoughmail.kubejs_tfc.event.*;
 import com.notenoughmail.kubejs_tfc.item.FluidContainerItemBuilder;
 import com.notenoughmail.kubejs_tfc.util.implementation.DataType;
 import com.notenoughmail.kubejs_tfc.util.implementation.KubeJSTFCCommands;
+import com.notenoughmail.kubejs_tfc.util.implementation.custom.climate.KubeJSClimateModel;
 import dev.latvian.mods.kubejs.bindings.event.PlayerEvents;
 import dev.latvian.mods.kubejs.event.EventGroup;
 import dev.latvian.mods.kubejs.event.EventHandler;
@@ -12,7 +13,10 @@ import dev.latvian.mods.kubejs.event.EventJS;
 import dev.latvian.mods.kubejs.script.data.DataPackEventJS;
 import net.dries007.tfc.util.DataManager;
 import net.dries007.tfc.util.DispenserBehaviors;
+import net.dries007.tfc.util.climate.Climate;
 import net.dries007.tfc.util.events.*;
+import net.dries007.tfc.world.ChunkGeneratorExtension;
+import net.dries007.tfc.world.chunkdata.ChunkData;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
@@ -20,9 +24,11 @@ import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.DispenserBlock;
+import net.minecraft.world.level.chunk.LevelChunk;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.entity.player.PlayerContainerEvent;
+import net.minecraftforge.event.level.ChunkEvent;
 import net.minecraftforge.event.server.ServerAboutToStartEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
@@ -77,6 +83,7 @@ public class EventHandlers {
         bus.addListener(EventHandlers::onDouseFire);
         bus.addListener(EventHandlers::serverAboutToStart);
         bus.addListener(KubeJSTFCCommands::reg);
+        bus.addListener(EventHandlers::onChunkLoad);
         if (!FMLEnvironment.production) {
             bus.addListener(EventPriority.LOWEST, EventHandlers::reloadListeners);
         }
@@ -227,5 +234,16 @@ public class EventHandlers {
                 consumer.accept(manager);
             }
         }).forEach(manager -> KubeJSTFC.warningLog("DataManager has not been handled: {} ({})", manager.directory, manager));
+    }
+
+    // TODO: Test/verify
+    private static void onChunkLoad(ChunkEvent.Load event) {
+        if (event.isNewChunk() && event.getChunk() instanceof LevelChunk lc && lc.getLevel() instanceof ServerLevel sl && !(sl.getChunkSource().getGenerator() instanceof ChunkGeneratorExtension) && Climate.model(sl) instanceof KubeJSClimateModel model) {
+            final ChunkData chunkData = new ChunkData(lc.getPos());
+            model.createChunkData(sl, lc, chunkData);
+            if (chunkData.status() == ChunkData.Status.FULL || chunkData.status() == ChunkData.Status.PARTIAL) {
+                ChunkData.update(lc, chunkData);
+            }
+        }
     }
 }