+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/custom/climate/KubeJSClimateModel.java
@@ -37,7 +37,7 @@ public class KubeJSClimateModel implements ClimateModel {
 
     private final Builder builder;
     private WindVectorCallback wind;
-    private OnChunkLoadCallback chunkLoad;
+    private OnChunkLoadCallback chunkLoad, createChunkData;
     private TemperatureCallback currentTemperature;
     private LevelPos2FloatCallback averageTemp, averageRain;
     private LevelPosLong2FloatCallback waterFog, airFog;
@@ -50,6 +50,7 @@ public class KubeJSClimateModel implements ClimateModel {
         this.builder = builder;
         wind = (level, pos, calendarTicks) -> Vec2.ZERO;
         chunkLoad = (level, chunk, chunkData) -> {};
+        createChunkData = (level, chunk, chunkData) -> {};
         currentTemperature = (level, pos, calendarTicks, daysInMonth) -> 0;
         averageTemp = (level, pos) -> 0;
         averageRain = (level, pos) -> 0;
@@ -121,6 +122,10 @@ public class KubeJSClimateModel implements ClimateModel {
     @Override
     public void onChunkLoad(WorldGenLevel level, ChunkAccess chunk, ChunkData chunkData) {
         chunkLoad.apply(level, chunk, chunkData);
+    }
+
+    public void createChunkData(WorldGenLevel level, ChunkAccess chunk, ChunkData chunkData) {
+        createChunkData.apply(level, chunk, chunkData);
     }
 
     @Override
@@ -270,6 +275,11 @@ public class KubeJSClimateModel implements ClimateModel {
         @Info(value = "Sets the model's behavior on chunk load")
         public void setOnChunkLoad(OnChunkLoadCallback callback) {
             model.chunkLoad = callback;
+        }
+
+        @Info(value = "Called on generation of a new chunk when the world generator is not TFC-like, enabling population of `ChunkData` values")
+        public void setCreateChunkData(OnChunkLoadCallback callback) {
+            model.createChunkData = callback;
         }
 
         @Info(value = "Adds a new Noise2D to the model, which can be retrieved later via the returned index")